@@ -1,7 +1,7 @@
 {
   "name": "allusion",
   "productName": "Allusion",
-  "version": "1.0.0-rc8.1",
+  "version": "1.0.0-rc9",
   "description": "A tool for managing your visual library",
   "main": "build/main.bundle.js",
   "scripts": {
@@ -82,20 +82,20 @@
   "homepage": "https://github.com/allusion-app/Allusion#readme",
   "devDependencies": {
     "@svgr/webpack": "^6.1.2",
-    "@types/chrome": "^0.0.86",
+    "@types/chrome": "^0.0.195",
     "@types/fs-extra": "^9.0.0",
     "@types/jest": "^27.4.0",
     "@types/offscreencanvas": "^2019.6.2",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
     "@types/react-window": "^1.8.5",
     "@types/utif": "^3.0.1",
     "@types/uuid": "^8.3.0",
     "@typescript-eslint/eslint-plugin": "^5.9.0",
     "@typescript-eslint/parser": "^5.9.0",
     "css-loader": "^6.5.1",
-    "electron": "18.0.0",
-    "electron-builder": "^22.14.13",
+    "electron": "20.0.3",
+    "electron-builder": "^23.3.3",
     "eslint": "^8.6.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
@@ -131,9 +131,9 @@
     "mobx-react-lite": "^3.2.3",
     "node-exiftool": "^2.3.0",
     "node-stream-zip": "^1.15.0",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-colorful": "^5.5.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-window": "^1.8.6",
     "sourcemapped-stacktrace": "^1.1.11",
     "utif": "^3.1.0",
